okay, generics for simplification?
