+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod bitcoin;
 pub mod config;
 pub mod context;
 pub mod deposit_monitor;
+pub mod dispatch;
 pub mod error;
 pub mod logging;
 pub mod reward_claim_process;
@@ -15,8 +16,8 @@ pub mod storage;
 pub mod testing;
 
 /// The capacity of the channel for sending new Bitcoin chain tips to a
-/// process. 
-/// 
+/// process.
+///
 /// Each process should have their own channel to avoid blocking the main
 /// broadcast channel.
 pub const MALBOX_CAPACITY: usize = 1024;
